Moved exception to its own namespace
