--- src/ModuleExceptionInterface.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Dhii\Modular\Module;
-
-/**
- * Represents an exception that is thrown in relation to a module.
- *
- * @since [*next-version*]
- */
-interface ModuleExceptionInterface extends
-    ModuleAwareInterface,
-    ModuleKeyAwareInterface
-{
-}
